merge

+++ src/link/multicast/udp.c
@@ -93,7 +93,7 @@ int8_t _z_endpoint_udp_multicast_valid(_z_endpoint_t *endpoint) {
         if (s_port == NULL) {
             ret = _Z_ERR_CONFIG_LOCATOR_INVALID;
         } else {
-            uint32_t port = strtoul(s_port, NULL, 10);
+            uint32_t port = (uint32_t)strtoul(s_port, NULL, 10);
             if ((port < (uint32_t)1) || (port > (uint32_t)65355)) {  // Port numbers should range from 1 to 65355
                 ret = _Z_ERR_CONFIG_LOCATOR_INVALID;
             }
@@ -115,7 +115,7 @@ int8_t _z_f_link_open_udp_multicast(_z_link_t *self) {
     uint32_t tout = Z_CONFIG_SOCKET_TIMEOUT;
     char *tout_as_str = _z_str_intmap_get(&self->_endpoint._config, UDP_CONFIG_TOUT_KEY);
     if (tout_as_str != NULL) {
-        tout = strtoul(tout_as_str, NULL, 10);
+        tout = (uint32_t)strtoul(tout_as_str, NULL, 10);
     }
 
     const char *iface = _z_str_intmap_get(&self->_endpoint._config, UDP_CONFIG_IFACE_KEY);
